adding the fixed file
The file i'm adding is the code to fix the bounce. What the issue is is that when the button is clicked fast, sometimes there is a bounce where it does not change the light color/colour. Now there is a delay added so that it waits long enough for the bounce to end and it can be ready for the new switch. the day is of 150, i believe millisecond.